Moves CustomerInfo and dependencies.
