+++ apiTester/src/commonMain/kotlin/com/revenuecat/purchases/kmp/apitester/PaywallListenerAPI.kt
@@ -1,8 +1,8 @@
 package com.revenuecat.purchases.kmp.apitester
 
-import com.revenuecat.purchases.kmp.CustomerInfo
 import com.revenuecat.purchases.kmp.Package
 import com.revenuecat.purchases.kmp.PurchasesError
+import com.revenuecat.purchases.kmp.models.CustomerInfo
 import com.revenuecat.purchases.kmp.models.StoreTransaction
 import com.revenuecat.purchases.kmp.ui.revenuecatui.PaywallListener
