+++ core/src/commonMain/kotlin/com/revenuecat/purchases/kmp/PurchasesConfiguration.kt
@@ -3,6 +3,7 @@ package com.revenuecat.purchases.kmp
 import com.revenuecat.purchases.kmp.Purchases.Companion.logHandler
 import com.revenuecat.purchases.kmp.PurchasesConfiguration.Builder
 import com.revenuecat.purchases.kmp.models.EntitlementVerificationMode
+import com.revenuecat.purchases.kmp.models.Store
 import com.revenuecat.purchases.kmp.models.StoreKitVersion
 import kotlin.jvm.JvmSynthetic
