+++ mappings/src/androidMain/kotlin/com/revenuecat/purchases/kmp/mappings/EntitlementInfos.android.kt
@@ -1,6 +1,6 @@
 package com.revenuecat.purchases.kmp.mappings
 
-import com.revenuecat.purchases.kmp.EntitlementInfos
+import com.revenuecat.purchases.kmp.models.EntitlementInfos
 import com.revenuecat.purchases.EntitlementInfos as RcEntitlementInfos
 
 internal fun RcEntitlementInfos.toEntitlementInfos(): EntitlementInfos =
