+++ models/src/commonMain/kotlin/com/revenuecat/purchases/kmp/models/EntitlementInfos.kt
@@ -1,6 +1,4 @@
-package com.revenuecat.purchases.kmp
-
-import com.revenuecat.purchases.kmp.models.VerificationResult
+package com.revenuecat.purchases.kmp.models
 
 /**
  * This class contains all the entitlements associated to the user.
